fix tabbing
BUG: 290959
BUG: 265160
BUG: 229292
BUG: 238279
BUG: 290758
BUG: 222831
BUG: 278275
BUG: 245747
BUG: 230000

BUG: 253697
BUG: 230570
BUG: 265977
BUG: 225337
BUG: 225339

REVIEW: 103855

+++ clients/oxygen/oxygenclientgroupitemdata.h
@@ -151,7 +151,7 @@ namespace Oxygen
         { return animation().data()->isRunning(); }
 
         //! update button activity
-        void updateButtonActivity( int visibleItem ) const;
+        void updateButtonActivity( long visibleItem ) const;
 
         //! update buttons
         void updateButtons( bool alsoUpdate ) const;
